+++ Source/Ardutosh/StringWrapper.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 
 class xString
 {
@@ -16,7 +17,7 @@ public:
 	int GetLineStartIndex(int lineNumber, int maxCharactersPerLine) const;
 	int GetIndex(char search, int start = 0) const;
 	int GetLineEndIndex(int start, int maxCharactersPerLine) const;
-	int Length() const { return CalculateLength(); }
+	size_t Length() const { return CalculateLength(); }
 	int GetWordLength(int index) const;
 	int NumLines(int maxCharactersPerLine) const;
 	const char* GetData() const { return data; }
@@ -26,7 +27,7 @@ public:
 
 private:
 	char Read(int i) const ;
-	int CalculateLength() const;
+	size_t CalculateLength() const;
 
 	char* data;
 	Type type;
